gritview.backend/requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.4.2
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
Werkzeug==1.0.1
